requirements.txt
Django
djangorestframework
psycopg2-binary
dj-database-url
djangorestframework-jwt
python-dotenv
djangorestframework-simplejwt==5.3.1
django-cors-headers
pydicom